+++ R/moran.R
@@ -16,7 +16,7 @@
 #' # Generating a small random graph
 #' set.seed(123)
 #' graph <- rgraph_ba(t = 4)
-#' w <- sna::geodist(as.matrix(w))
+#' w <- igraph::distances(igraph::graph_from_adjacency_matrix(graph))
 #' x <- rnorm(5)
 #'
 #' # Computing Moran's I