+++ src/stats.h
@@ -11,9 +11,13 @@ arma::colvec degree_cpp(
     const arma::sp_mat & adjmat0, const int & cmode=2,
     bool undirected=true, bool self=false, bool valued=false);
 
-arma::mat exposure_cpp(
-    List graph, arma::mat cumadopt, arma::mat attrs,
-    bool outgoing =true, bool valued=true, bool normalized=true);
+arma::colvec exposure_cpp(
+    const arma::sp_mat & graph,
+    const arma::colvec & cumadopt,
+    const arma::colvec & attrs,
+    bool outgoing = true,
+    bool valued = true,
+    bool normalized = true);
 
 arma::mat cumulative_adopt_count_cpp(const arma::mat & cumadopt);